+++ components/st-page-title/st-page-title.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, Output, ChangeDetectionStrategy, EventEmitter } from '@angular/core';
+import { Component, Input, Output, ChangeDetectionStrategy, EventEmitter, OnInit } from '@angular/core';
 
 @Component({
   selector: 'st-page-title',
@@ -6,13 +6,24 @@ import { Component, Input, Output, ChangeDetectionStrategy, EventEmitter } from 
   styleUrls: ['./st-page-title.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class StPageTitleComponent {
+export class StPageTitleComponent implements OnInit {
    @Input() title: string = '';
    @Input() leftButton: string = '';
-   @Input() qaTag: string = '';
+   @Input() qaTag: string;
+   @Input() preTitle: string | undefined;
    @Output() clickButton: EventEmitter<any> = new EventEmitter();
-   
+
    onClickedButton(): void {
       this.clickButton.emit();
+   }
+
+   hasPreTitle(): boolean {
+      return this.preTitle !== undefined;
+   }
+
+   ngOnInit(): void {
+      if (this.qaTag === undefined) {
+         throw new Error('ST-PAGE-TITLE: qa field is required');
+      }
    }
 }