+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Remove st-spinner from website
 * Add Button Clear in Search Component
 * Add Value Input in Search Component
+* Add optional Pre-title to Page Title Component
 * BREAKING CHANGE. Remove ng2-translate dependency
 * BREAKING CHANGE. Now egeo is imported like this: "import { EgeoModule } from 'egeo';"
 * BREAKING CHANGE. Output Search Component change of (onSearch) to (search);